requirements.txt
alibabacloud-credentials==0.1.4
alibabacloud-dysmsapi20170525==2.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.4
alibabacloud-tea==0.2.7
alibabacloud-tea-openapi==0.2.5
alibabacloud-tea-util==0.3.4
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==3.2.5
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
idna==2.10
kombu==5.1.0
Markdown==3.3.4
mysqlclient==2.0.3
prompt-toolkit==3.0.19
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
